rename `model` package to `models`
