+++ app/shared/src/androidMain/kotlin/ui/subject/collection/EpisodeProgress.android.kt
@@ -4,7 +4,7 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import androidx.compose.ui.tooling.preview.PreviewLightDark
-import me.him188.ani.app.data.model.PackedDate
+import me.him188.ani.app.data.models.PackedDate
 import me.him188.ani.app.data.source.media.EpisodeCacheStatus
 import me.him188.ani.app.ui.cache.testMediaCache1
 import me.him188.ani.app.ui.foundation.ProvideCompositionLocalsForPreview
