+++ app/shared/src/androidMain/kotlin/ui/subject/details/SubjectDetailsPage.android.kt
@@ -7,13 +7,13 @@ import androidx.compose.ui.input.nestedscroll.nestedScroll
 import androidx.compose.ui.tooling.preview.Devices
 import androidx.compose.ui.tooling.preview.Preview
 import kotlinx.coroutines.flow.MutableStateFlow
-import me.him188.ani.app.data.model.subject.CharacterType
-import me.him188.ani.app.data.model.subject.Images
-import me.him188.ani.app.data.model.subject.PersonCareer
-import me.him188.ani.app.data.model.subject.PersonInfo
-import me.him188.ani.app.data.model.subject.PersonType
-import me.him188.ani.app.data.model.subject.RelatedCharacterInfo
-import me.him188.ani.app.data.model.subject.RelatedPersonInfo
+import me.him188.ani.app.data.models.subject.CharacterType
+import me.him188.ani.app.data.models.subject.Images
+import me.him188.ani.app.data.models.subject.PersonCareer
+import me.him188.ani.app.data.models.subject.PersonInfo
+import me.him188.ani.app.data.models.subject.PersonType
+import me.him188.ani.app.data.models.subject.RelatedCharacterInfo
+import me.him188.ani.app.data.models.subject.RelatedPersonInfo
 import me.him188.ani.app.ui.foundation.ProvideCompositionLocalsForPreview
 import me.him188.ani.app.ui.foundation.layout.rememberConnectedScrollState
 import me.him188.ani.app.ui.subject.collection.TestSelfRatingInfo
