+++ app/shared/src/androidMain/kotlin/ui/subject/search/SubjectList.preview.kt
@@ -21,8 +21,8 @@ package me.him188.ani.app.ui.subject.search
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import androidx.compose.ui.tooling.preview.Preview
-import me.him188.ani.app.data.model.subject.RatingInfo
-import me.him188.ani.app.data.model.subject.SubjectInfo
+import me.him188.ani.app.data.models.subject.RatingInfo
+import me.him188.ani.app.data.models.subject.SubjectInfo
 import me.him188.ani.app.ui.foundation.ProvideCompositionLocalsForPreview
 import me.him188.ani.app.ui.home.search.SubjectListViewModel
 import me.him188.ani.app.ui.home.search.SubjectPreviewColumn
