+++ app/shared/src/commonMain/kotlin/data/models/subject/RatingInfo.kt
@@ -1,4 +1,4 @@
-package me.him188.ani.app.data.model.subject
+package me.him188.ani.app.data.models.subject
 
 import androidx.compose.runtime.Immutable
 import androidx.compose.runtime.Stable
