+++ app/shared/src/commonMain/kotlin/data/models/subject/RelatedPersonInfo.kt
@@ -1,4 +1,4 @@
-package me.him188.ani.app.data.model.subject
+package me.him188.ani.app.data.models.subject
 
 class RelatedPersonInfo(
     val personInfo: PersonInfo,
