+++ app/shared/src/commonMain/kotlin/data/models/subject/SubjectAiringInfo.kt
@@ -1,13 +1,13 @@
-package me.him188.ani.app.data.model.subject
+package me.him188.ani.app.data.models.subject
 
 import androidx.compose.runtime.Immutable
 import androidx.compose.runtime.Stable
-import me.him188.ani.app.data.model.PackedDate
-import me.him188.ani.app.data.model.episode.EpisodeInfo
-import me.him188.ani.app.data.model.episode.isKnownCompleted
-import me.him188.ani.app.data.model.episode.isKnownOnAir
-import me.him188.ani.app.data.model.ifInvalid
-import me.him188.ani.app.data.model.minus
+import me.him188.ani.app.data.models.PackedDate
+import me.him188.ani.app.data.models.episode.EpisodeInfo
+import me.him188.ani.app.data.models.episode.isKnownCompleted
+import me.him188.ani.app.data.models.episode.isKnownOnAir
+import me.him188.ani.app.data.models.ifInvalid
+import me.him188.ani.app.data.models.minus
 import me.him188.ani.datasources.api.EpisodeSort
 import kotlin.time.Duration.Companion.days
 import kotlin.time.times
