+++ app/shared/src/commonMain/kotlin/data/models/subject/SubjectInfo.kt
@@ -1,9 +1,9 @@
-package me.him188.ani.app.data.model.subject
+package me.him188.ani.app.data.models.subject
 
 import androidx.compose.runtime.Immutable
 import androidx.compose.runtime.Stable
 import kotlinx.serialization.Serializable
-import me.him188.ani.app.data.model.PackedDate
+import me.him188.ani.app.data.models.PackedDate
 import me.him188.ani.utils.logging.info
 import me.him188.ani.utils.logging.logger
