+++ app/shared/src/commonMain/kotlin/data/repository/BangumiRelatedCharactersRepository.kt
@@ -9,13 +9,13 @@ import kotlinx.coroutines.withContext
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.jsonObject
-import me.him188.ani.app.data.model.subject.CharacterType
-import me.him188.ani.app.data.model.subject.Images
-import me.him188.ani.app.data.model.subject.PersonCareer
-import me.him188.ani.app.data.model.subject.PersonInfo
-import me.him188.ani.app.data.model.subject.PersonType
-import me.him188.ani.app.data.model.subject.RelatedCharacterInfo
-import me.him188.ani.app.data.model.subject.RelatedPersonInfo
+import me.him188.ani.app.data.models.subject.CharacterType
+import me.him188.ani.app.data.models.subject.Images
+import me.him188.ani.app.data.models.subject.PersonCareer
+import me.him188.ani.app.data.models.subject.PersonInfo
+import me.him188.ani.app.data.models.subject.PersonType
+import me.him188.ani.app.data.models.subject.RelatedCharacterInfo
+import me.him188.ani.app.data.models.subject.RelatedPersonInfo
 import me.him188.ani.datasources.bangumi.BangumiClient
 import me.him188.ani.datasources.bangumi.models.BangumiCharacterType
 import me.him188.ani.datasources.bangumi.models.BangumiPerson
