+++ app/shared/src/commonMain/kotlin/ui/home/search/SubjectListViewModel.kt
@@ -23,7 +23,7 @@ import kotlinx.coroutines.CoroutineStart
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.yield
-import me.him188.ani.app.data.model.subject.SubjectInfo
+import me.him188.ani.app.data.models.subject.SubjectInfo
 import me.him188.ani.app.ui.foundation.AbstractViewModel
 import me.him188.ani.app.ui.foundation.launchInBackground
 import me.him188.ani.datasources.api.paging.PagedSource
