+++ app/shared/src/commonMain/kotlin/ui/subject/collection/CollectionPage.kt
@@ -64,8 +64,8 @@ import kotlinx.coroutines.CancellationException
 import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.launch
-import me.him188.ani.app.data.model.episode.episode
-import me.him188.ani.app.data.model.subject.SubjectCollection
+import me.him188.ani.app.data.models.episode.episode
+import me.him188.ani.app.data.models.subject.SubjectCollection
 import me.him188.ani.app.navigation.LocalNavigator
 import me.him188.ani.app.platform.Platform
 import me.him188.ani.app.platform.currentPlatform
