+++ app/shared/src/commonMain/kotlin/ui/subject/collection/MyCollectionColumn.kt
@@ -58,14 +58,14 @@ import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.coerceAtLeast
 import androidx.compose.ui.unit.dp
 import kotlinx.coroutines.flow.debounce
-import me.him188.ani.app.data.model.episode.EpisodeCollection
-import me.him188.ani.app.data.model.episode.episode
-import me.him188.ani.app.data.model.subject.SubjectAiringInfo
-import me.him188.ani.app.data.model.subject.SubjectAiringKind
-import me.him188.ani.app.data.model.subject.SubjectCollection
-import me.him188.ani.app.data.model.subject.getEpisodeToPlay
-import me.him188.ani.app.data.model.subject.isOnAir
-import me.him188.ani.app.data.model.toStringExcludingSameYear
+import me.him188.ani.app.data.models.episode.EpisodeCollection
+import me.him188.ani.app.data.models.episode.episode
+import me.him188.ani.app.data.models.subject.SubjectAiringInfo
+import me.him188.ani.app.data.models.subject.SubjectAiringKind
+import me.him188.ani.app.data.models.subject.SubjectCollection
+import me.him188.ani.app.data.models.subject.getEpisodeToPlay
+import me.him188.ani.app.data.models.subject.isOnAir
+import me.him188.ani.app.data.models.toStringExcludingSameYear
 import me.him188.ani.app.data.source.media.EpisodeCacheStatus
 import me.him188.ani.app.tools.caching.LazyDataCache
 import me.him188.ani.app.ui.foundation.AsyncImage
